feat(users): make avatar optional during registration

+++ apps/users/serializers.py
@@ -18,9 +18,7 @@ User = get_user_model()
 
 class RegisterSerializer(serializers.ModelSerializer):
     password: serializers.CharField = serializers.CharField(write_only=True, min_length=8)
-    avatar: serializers.ImageField = serializers.ImageField(
-        required=True, allow_empty_file=False
-    )
+    avatar = serializers.ImageField(required=False, allow_empty_file=True)
 
     @extend_schema_field({"type": "string", "format": "binary"})
     def get_avatar(self, obj: Any) -> None:
@@ -47,44 +45,48 @@ class RegisterSerializer(serializers.ModelSerializer):
 
         return value
 
-    def validate_avatar(self, value: Any) -> Any:
+    def validate_avatar(self, value):
+        if value is None:
+            return value
+        
         if isinstance(value, str):
             raise serializers.ValidationError(
                 _("Avatar must be uploaded as a file (multipart/form-data).")
             )
-
-        if not value:
-            raise serializers.ValidationError(_("Avatar is required."))
-
+        
         is_valid, error = AvatarProcessor.validate_image(value)
         if not is_valid:
             raise serializers.ValidationError(_(error))
-
+        
         return value
 
-    def create(self, validated_data: dict[str, Any]) -> Any:
-        password: str = validated_data.pop("password")
-        avatar_file = validated_data.pop("avatar")
+    def create(self, validated_data):
+        password = validated_data.pop("password")
+        avatar_file = validated_data.pop("avatar", None)
         user = User.objects.create_user(password=password, **validated_data)
 
-        firestore_service = FirestoreUserService()
-
-        base64_avatar, error = AvatarProcessor.process_avatar(avatar_file)
-        if error:
-            user.delete()
-            raise serializers.ValidationError({"avatar": error})
-
-        avatar_id: str | None = firestore_service.create_user_avatar(user.id, base64_avatar)
-
-        if avatar_id:
-            user.firestore_avatar_id = avatar_id
-            user.use_firestore_avatar = True
-            user.save(update_fields=["firestore_avatar_id", "use_firestore_avatar"])
+        if avatar_file:
+            firestore_service = FirestoreUserService()
+            base64_avatar, error = AvatarProcessor.process_avatar(avatar_file)
+            
+            if error:
+                user.delete()
+                raise serializers.ValidationError({"avatar": error})
+            
+            avatar_id = firestore_service.create_user_avatar(user.id, base64_avatar)
+            
+            if avatar_id:
+                user.firestore_avatar_id = avatar_id
+                user.use_firestore_avatar = True
+                user.save(update_fields=["firestore_avatar_id", "use_firestore_avatar"])
+            else:
+                user.delete()
+                raise serializers.ValidationError(
+                    {"avatar": "Failed to store avatar. Please try again."}
+                )
         else:
-            user.delete()
-            raise serializers.ValidationError(
-                {"avatar": "Failed to store avatar. Please try again."}
-            )
+            user.use_firestore_avatar = False
+            user.save(update_fields=["use_firestore_avatar"])
 
         return user
